토큰 만료시 로그아웃 처리

+++ client/src/pages/post/index.js
@@ -19,7 +19,9 @@ const Post = () => {
   };
 
   const onClickDeleteButton = () => {
-    dispatch(actions.deletePost({ id }));
+    if (window.confirm('정말로 삭제하시겠습니까?')) {
+      dispatch(actions.deletePost({ id }));
+    }
   };
 
   useEffect(() => {
